Extract dummy domain resolution
